# [release-1.10] Subtype: some performance tuning. (#56007)